+++ game/svg.cc
@@ -13,7 +13,7 @@
 #endif
 
 void loadSVG(Bitmap& bitmap, fs::path const& filename) {
-	double factor = config["graphic/svg_lod"].f();
+	float factor = config["graphic/svg_lod"].f();
 	// Try to load a cached PNG instead
 	if (cache::loadSVG(bitmap, filename, factor)) return;
 	std::clog << "image/debug: Loading SVG: " + filename.string() << std::endl;
